+++ praetorian_cli/main.py
@@ -1,6 +1,7 @@
 import click
 
 import praetorian_cli.handlers.add
+import praetorian_cli.handlers.agent
 import praetorian_cli.handlers.delete
 import praetorian_cli.handlers.enrich
 import praetorian_cli.handlers.get