+++ praetorian_cli/sdk/test/test_file.py
@@ -18,13 +18,16 @@ class TestFile:
         with open(self.local_filepath, 'w') as file:
             file.write(self.content)
 
+    def test_sanitized_filepath(self):
+        assert self.sdk.files.sanitize_filename(self.chariot_filepath) == self.sanitized_filepath
+
     def test_add_file(self):
         self.sdk.files.add(self.local_filepath, self.chariot_filepath)
         files, offset = self.sdk.files.list(self.chariot_filepath)
         assert files[0]['name'] == self.chariot_filepath
 
     def test_get_file(self):
-        self.sdk.files.get(self.chariot_filepath, os.getcwd())
+        self.sdk.files.save(self.chariot_filepath, os.getcwd())
         with open(self.sanitized_filepath, 'r') as f:
             assert f.read() == self.content